+++ src/formats/csv/CsvFormatDefinition.ts
@@ -1,0 +1,89 @@
+import { parse, unparse } from 'papaparse';
+import spaceTrim from 'spacetrim';
+import { Parameters } from '../../_packages/types.index';
+import { ParseError } from '../../errors/ParseError';
+import { TODO_USE } from '../../utils/organization/TODO_USE';
+import type { FormatDefinition } from '../_common/FormatDefinition';
+
+/**
+ * Definition for CSV spreadsheet
+ *
+ * @private still in development [🏢]
+ */
+export const CsvFormatDefinition: FormatDefinition<string /* <- [0] */, string /* <- [👨‍⚖️] */, object /* <- [1] */> = {
+    formatName: 'CSV',
+
+    aliases: ['SPREADSHEET', 'TABLE'],
+
+    isValid(value, schema): value is string /* <- [0] */ {
+        TODO_USE(value /* <- TODO: Use value here */);
+        TODO_USE(schema /* <- TODO: Use schema here */);
+        return true;
+    },
+
+    canBeValid(partialValue, schema): partialValue is string /* <- [0] */ {
+        TODO_USE(partialValue /* <- TODO: Use partialValue here */);
+        TODO_USE(schema /* <- TODO: Use schema here */);
+        return true;
+    },
+
+    heal(value, schema) {
+        TODO_USE(value /* <- TODO: Use partialValue here */);
+        TODO_USE(schema /* <- TODO: Use schema here */);
+        throw new Error('Not implemented');
+    },
+
+    subvalueDefinitions: [
+        {
+            subvalueName: 'ROW',
+            async mapValues(value, mapCallback) {
+                // TODO: [👨🏾‍🤝‍👨🏼] DRY csv parsing
+                const csv = parse(value, {
+                    header: true,
+                    delimiter: ',',
+                    quoteChar: '"',
+                    newline: '\r\n',
+                    skipEmptyLines: true,
+                    // encoding: 'utf8',
+                    // <- TODO: !!!!!! DEFAULT_CSV_OPTIONS
+                    // <- TODO: [🧠] How to define parsing options for formats, its different concept than schema
+                });
+
+                if (csv.errors.length !== 0) {
+                    throw new ParseError( // <- TODO: !!!!!! Split PipelineParseError and FormatParseError -> CsvParseError
+                        spaceTrim(
+                            (block) => `
+                                CSV parsing error
+
+                                ${block(csv.errors.map((error) => error.message).join('\n\n'))}
+                            `,
+                        ),
+                    );
+                }
+
+                const mappedData = await Promise.all(
+                    csv.data.map((row, index) => /*not await */ mapCallback(row as Parameters, index)),
+                );
+
+                return unparse(mappedData, {
+                    header: true,
+                    delimiter: ',',
+                    quoteChar: '"',
+                    newline: '\r\n',
+                    skipEmptyLines: true,
+                    // encoding: 'utf8',
+                    // <- TODO: !!!!!! DEFAULT_CSV_OPTIONS
+                    // <- TODO: [🧠] How to define parsing options for formats, its different concept than schema
+                });
+            },
+        },
+    ],
+};
+
+/**
+ * TODO: [🍓] In `CsvFormatDefinition` implement simple `isValid`
+ * TODO: [🍓] In `CsvFormatDefinition` implement partial `canBeValid`
+ * TODO: [🍓] In `CsvFormatDefinition` implement `heal
+ * TODO: [🍓] In `CsvFormatDefinition` implement `subvalueDefinitions`
+ * TODO: [🏢] Allow to expect something inside CSV objects and other formats
+ */