+++ src/formats/json/JsonFormatDefinition.ts
@@ -44,6 +44,6 @@ export const JsonFormatDefinition: FormatDefinition<string /* <- [0] */, string 
  * TODO: [🍓] In `JsonFormatDefinition` implement simple `isValid`
  * TODO: [🍓] In `JsonFormatDefinition` implement partial `canBeValid`
  * TODO: [🍓] In `JsonFormatDefinition` implement `heal
- * TODO: [🍓] In `JsonFormatDefinition` implement `extractValues`
+ * TODO: [🍓] In `JsonFormatDefinition` implement `subvalueDefinitions`
  * TODO: [🏢] Allow to expect something inside JSON objects and other formats
  */